要命 力控又有点问题（目前猜测是发送的问题）

+++ SB_Up/SB_Up/User/PotocaLayer/MF_User.h
@@ -10,7 +10,7 @@
 void Start_MF_send(int16_t ID);
 void Current_Control_MF_send(int16_t ID,int16_t iqControl);
 int16_t Current_Limit_MF(int16_t current);
-float MF_value(int16_t k, int16_t n, int16_t max);
+float MF_value(int32_t k, int32_t n, int32_t max);
 void Encoder_MF_read(int16_t ID);
 	 
 #ifdef __cplusplus
